+++ android/src/main/kotlin/ru/grfc/telephony/flutter_telephony_info/FlutterTelephonyInfoPlugin.kt
@@ -8,47 +8,51 @@ import androidx.annotation.RequiresApi
 import io.flutter.embedding.engine.plugins.FlutterPlugin
 
 /** PigeonPlugin */
-class FlutterTelephonyInfoPlugin : FlutterPlugin, Telephony.TelephonyAPI {
+class FlutterTelephonyInfoPlugin : FlutterPlugin, TelephonyAPI {
     private lateinit var context: Context
     private lateinit var mTelephonyManager: TelephonyManager
 
     override fun onAttachedToEngine(flutterPluginBinding: FlutterPlugin.FlutterPluginBinding) {
-        Telephony.TelephonyAPI.setup(flutterPluginBinding.binaryMessenger, this)
+        TelephonyAPI.setUp(flutterPluginBinding.binaryMessenger, this)
         context = flutterPluginBinding.applicationContext
     }
 
     override fun onDetachedFromEngine(binding: FlutterPlugin.FlutterPluginBinding) {
-        Telephony.TelephonyAPI.setup(binding.binaryMessenger, null)
+        TelephonyAPI.setUp(binding.binaryMessenger, null)
     }
 
     @SuppressLint("MissingPermission")
-    @RequiresApi(Build.VERSION_CODES.R)
-    override fun getInfo(): List<Telephony.TelephonyInfo> {
+    override fun getInfo(): List<TelephonyInfo> {
         mTelephonyManager = context.getSystemService(Context.TELEPHONY_SERVICE) as TelephonyManager
 
-        val telephonyList: MutableList<Telephony.TelephonyInfo> = mutableListOf()
+        var cellId: String
+        var cellSignalStrength: String
+        val telephonyList: MutableList<TelephonyInfo> = mutableListOf()
 
         for (i in 0 until mTelephonyManager.activeModemCount) {
-            val telephonyInfo = Telephony.TelephonyInfo()
-            telephonyInfo.displayName = mTelephonyManager.simOperatorName
-            telephonyInfo.mobileNetworkCode = mTelephonyManager.simOperator
-            telephonyInfo.radioType = radioType()
-            telephonyInfo.networkGeneration = networkGeneration()
             if (mTelephonyManager.allCellInfo.isEmpty()) {
-                telephonyInfo.cellId = "Error! GPS must be turned on."
-                telephonyInfo.cellSignalStrength = "Error! GPS must be turned on."
+                cellId = "Error! GPS must be turned on."
+                cellSignalStrength = "Error! GPS must be turned on."
             } else {
-                telephonyInfo.cellId = mTelephonyManager.allCellInfo[i].cellIdentity.toString()
-                telephonyInfo.cellSignalStrength =
+                cellId = mTelephonyManager.allCellInfo[i].cellIdentity.toString()
+                cellSignalStrength =
                     mTelephonyManager.allCellInfo[i].cellSignalStrength.toString()
             }
+            val telephonyInfo = TelephonyInfo(
+                displayName = mTelephonyManager.simOperatorName,
+                mobileNetworkCode = mTelephonyManager.simOperator,
+                radioType = radioType(),
+                networkGeneration = networkGeneration(),
+                cellId = cellId,
+                cellSignalStrength = cellSignalStrength
+            )
+
             telephonyList += telephonyInfo
         }
         return telephonyList
     }
 
     @SuppressLint("MissingPermission")
-    @RequiresApi(Build.VERSION_CODES.N)
     private fun radioType(): String {
         return when (mTelephonyManager.dataNetworkType) {
             TelephonyManager.NETWORK_TYPE_1xRTT -> return "1xRTT"
@@ -86,6 +90,7 @@ class FlutterTelephonyInfoPlugin : FlutterPlugin, Telephony.TelephonyAPI {
             TelephonyManager.NETWORK_TYPE_IDEN,
             TelephonyManager.NETWORK_TYPE_GSM
             -> return "2G"
+
             TelephonyManager.NETWORK_TYPE_UMTS,
             TelephonyManager.NETWORK_TYPE_EVDO_0,
             TelephonyManager.NETWORK_TYPE_EVDO_A,
@@ -97,10 +102,13 @@ class FlutterTelephonyInfoPlugin : FlutterPlugin, Telephony.TelephonyAPI {
             TelephonyManager.NETWORK_TYPE_HSPAP,
             TelephonyManager.NETWORK_TYPE_TD_SCDMA
             -> return "3G"
+
             TelephonyManager.NETWORK_TYPE_LTE
             -> return "4G"
+
             TelephonyManager.NETWORK_TYPE_NR,
             -> return "5G"
+
             else -> radioType.toString()
         }
         return "unknown"