+++ CHANGELOG.md
@@ -1,5 +1,11 @@
-## [0.0.1-dev] - 2023-03-26
-- [ ADD ] Initial code;
+## [0.1.0] - 2023-05-14
+- [ CHANGE ] Dart 3.0.0 ready;
+- [ CHANGE ] The Pigeon API generator has been changed from Java to Kotlin;
+- [ CHANGE ] License;
+- [ ADD ] Pigeon config in code;
 
 ## [0.0.1] - 2023-03-26
 - [ CHANGE ] Change SDK version to '>=2.17.0 <4.0.0';
+
+## [0.0.1-dev] - 2023-03-26
+- [ ADD ] Initial code;